Sync backend/ to server state: hardening, sync fixes, rate limiting
Brings the monorepo copy of the backend up to the deployed codebase:

- Auth: rate-limited login/register/google (429 + Retry-After), new
  POST /auth/change-password (revokes all sessions, returns fresh pair)
  and PUT /auth/me; bcrypt + Google token verification moved off the
  event loop; Google sign-in requires a verified email before linking
  or creating an account.
- PUT contract: omitted field = unchanged; explicit null/"" on
  notebook_id/parent_id detaches (model_fields_set).
- Sync: notebook/parent ownership validated on push (foreign, unknown,
  malformed or self-parent ids degrade to unfiled/top-level instead of
  cross-tenant refs); deleted tag names no longer leak through pull;
  cycle-trapped notebooks always surface in the tree.
- Import/export: .oblix round-trip and .enex import (transfer service,
  upload policy), unicode filenames, extensionless uploads.
- Note versions coalesced+capped; noload/expire pattern kills the
  list/sync over-fetch; edited_at drives LWW (new migrations 0003/0004).
- Docs: backend README, CLAUDE.md, DEPLOY.md nginx X-Forwarded-For.

Verified by a 66-check live HTTP audit plus the 44-check smoke suite
against a production-mode boot (Alembic-migrated Postgres): 110/110.

+++ backend/app/models/file.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime
-from sqlalchemy import String, Integer, Boolean, DateTime, ForeignKey, func
+from sqlalchemy import String, BigInteger, Boolean, DateTime, ForeignKey, func
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.database import Base
@@ -15,7 +15,7 @@ class File(Base):
     filename: Mapped[str] = mapped_column(String(500), nullable=False)
     original_name: Mapped[str] = mapped_column(String(500), nullable=False)
     mime_type: Mapped[str] = mapped_column(String(255), nullable=False)
-    size_bytes: Mapped[int] = mapped_column(Integer, nullable=False)
+    size_bytes: Mapped[int] = mapped_column(BigInteger, nullable=False)
     storage_path: Mapped[str] = mapped_column(String(1000), nullable=False)
     is_deleted: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False, index=True)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
